Reorganize generators around env --> type format
