--- packages/nx-plugin/src/generators/lib-ts/files/vite.config.ts__tmpl__
@@ -1,3 +1,0 @@
-import getConfig from '@eternagame/vite-utils';
-
-export default getConfig({ type: 'lib', env: '<%= env %>' });
