+++ packages/nx-plugin/src/generators/package/base/files/README.md__tmpl__
