+++ packages/nx-plugin/src/generators/package/base/files/package.json__tmpl__
