+++ packages/nx-plugin/src/generators/package/node/app/files/src/index.ts__tmpl__
