+++ packages/nx-plugin/src/generators/package/node/app/files/vite.config.ts__tmpl__
