+++ packages/nx-plugin/src/generators/package/node/lib/files/vite.config.ts__tmpl__
@@ -1,0 +1,3 @@
+import getConfig from '@eternagame/vite-utils';
+
+export default getConfig({ type: 'lib', env: 'node' });
