+++ packages/nx-plugin/src/generators/package/ts/lib/files/jest.config.js__tmpl__
