+++ packages/nx-plugin/src/generators/package/ts/lib/files/src/index.spec.ts__tmpl__
