+++ packages/nx-plugin/src/generators/package/ts/lib/files/src/index.ts__tmpl__
