+++ packages/nx-plugin/src/generators/package/ts/lib/files/tsconfig.build.json__tmpl__
