+++ packages/nx-plugin/src/generators/package/ts/lib/files/tsconfig.json__tmpl__
