+++ packages/nx-plugin/src/generators/package/ts/lib/files/tsconfig.spec.json__tmpl__
