+++ packages/nx-plugin/src/generators/package/ts/lib/index.ts
@@ -7,12 +7,11 @@ import {
   updateJson,
   type Tree,
 } from '@nrwl/devkit';
-import generatePackage from '../package';
-import { installDevDependencies } from '../../utils/dependencies';
+import generatePackage from '../../base';
+import { installDevDependencies } from '../../../../utils/dependencies';
 
 interface Schema {
   name: string;
-  env: 'iso' | 'web' | 'node';
   description: string;
   directory: string;
 }
@@ -79,53 +78,12 @@ function updatePackageJson(tree: Tree, options: NormalizedSchema) {
   /* eslint-enable no-param-reassign */
 }
 
-function updateTsconfigs(tree: Tree, options: NormalizedSchema) {
-  /* eslint-disable no-param-reassign */
-  if (options.env === 'web') {
-    updateJson(
-      tree,
-      path.join(options.projectRoot, 'tsconfig.build.json'),
-      (json: Record<string, unknown>) => {
-        json['extends'] = '@eternagame/tsconfig/tsconfig.web.json';
-        return json;
-      },
-    );
-    updateJson(
-      tree,
-      path.join(options.projectRoot, 'tsconfig.spec.json'),
-      (json: Record<string, unknown>) => {
-        json['extends'] = '@eternagame/tsconfig/tsconfig.jest-web.json';
-        return json;
-      },
-    );
-  } else if (options.env === 'node') {
-    updateJson(
-      tree,
-      path.join(options.projectRoot, 'tsconfig.build.json'),
-      (json: Record<string, unknown>) => {
-        json['extends'] = '@eternagame/tsconfig/tsconfig.node.json';
-        return json;
-      },
-    );
-    updateJson(
-      tree,
-      path.join(options.projectRoot, 'tsconfig.spec.json'),
-      (json: Record<string, unknown>) => {
-        json['extends'] = '@eternagame/tsconfig/tsconfig.jest-node.json';
-        return json;
-      },
-    );
-  }
-  /* eslint-enable no-param-reassign */
-}
-
 export default async function generate(tree: Tree, options: Schema) {
   const normalizedOptions = normalizeOptions(tree, options);
 
   const finalizePackage = await generatePackage(tree, options);
   addFiles(tree, normalizedOptions);
   updatePackageJson(tree, normalizedOptions);
-  updateTsconfigs(tree, normalizedOptions);
 
   return async () => {
     await finalizePackage();
