+++ packages/nx-plugin/src/generators/package/vue/app/files/src/App.vue
