+++ packages/nx-plugin/src/generators/package/vue/app/files/src/components/WelcomeItem.vue
