+++ packages/nx-plugin/src/generators/package/vue/app/files/src/components/icons/IconSupport.vue
