+++ packages/nx-plugin/src/generators/package/vue/app/files/src/index.ts__tmpl__
