+++ packages/nx-plugin/src/generators/package/vue/app/files/src/shims-vue.d.ts
