+++ packages/nx-plugin/src/generators/package/vue/app/files/vite.config.ts__tmpl__
