+++ packages/nx-plugin/src/generators/package/vue/app/schema.json
@@ -1,8 +1,8 @@
 {
   "$schema": "http://json-schema.org/schema",
-  "$id": "AppNode",
+  "$id": "vue-app",
   "cli": "nx",
-  "title": "Adds a new Typescript NodeJS application to an Nx project",
+  "title": "Adds a new Vue browser application to an Nx project",
   "type": "object",
   "properties": {
     "name": {
