+++ packages/nx-plugin/src/generators/package/web/app/files/src/index.ts__tmpl__
