+++ packages/nx-plugin/src/generators/package/web/app/files/src/vite-env.d.ts
