+++ packages/nx-plugin/src/generators/package/web/app/files/vite.config.ts__tmpl__
