+++ packages/nx-plugin/src/generators/package/web/app/schema.json
@@ -1,0 +1,34 @@
+{
+  "$schema": "http://json-schema.org/schema",
+  "$id": "web-app",
+  "cli": "nx",
+  "title": "Adds a new Typescript browser application to an Nx project",
+  "type": "object",
+  "properties": {
+    "name": {
+      "type": "string",
+      "description": "Package name",
+      "$default": {
+        "$source": "argv",
+        "index": 0
+      },
+      "x-prompt": "What name of your npm package?",
+      "pattern": "^[a-zA-Z].*$"
+    },
+    "description": {
+      "type": "string",
+      "description": "Package description",
+      "$default": {
+        "$source": "argv",
+        "index": 1
+      },
+      "x-prompt": "Provide a brief description for the package",
+      "pattern": "^[\\w -_']*$"
+    },
+    "directory": {
+      "type": "string",
+      "description": "A directory under the packages directory to create the package in"
+    }
+  },
+  "required": ["name", "description"]
+}
